Move the `markdown` unit tests to the integration tests
The `sections` unit tests only exercise the public API (`parse`,
`sections`, `Item`, and `Content`), so they move verbatim from the
`#[cfg(test)]` module in `markdown.rs` to `tests/markdown.rs`,
joining the convergence tests.
